backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
mysql-connector-python
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
email-validator